requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-binance==1.0.19
coinbase==2.1.0
krakenex==2.1.0
stripe==7.11.0
plotly==5.18.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
cryptography==42.0.2
APScheduler==3.10.4
requests==2.31.0
websocket-client==1.7.0
qrcode==7.4.2
Pillow==10.2.0
reportlab==4.1.0
python-dateutil==2.8.2
psutil==5.9.8
ccxt==4.2.15

